Update usage of netty-specific classes in GrpcTransportBuilder
