+++ core/src/test/java/tech/ydb/core/impl/pool/GrpcChannelTest.java
@@ -2,15 +2,11 @@ package tech.ydb.core.impl.pool;
 
 
 import io.grpc.ConnectivityState;
-import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
 
-import tech.ydb.core.grpc.GrpcTransport;
 import tech.ydb.core.grpc.GrpcTransportBuilder;
 
 
@@ -19,22 +15,13 @@ import tech.ydb.core.grpc.GrpcTransportBuilder;
  * @author Aleksandr Gorshenin
  */
 public class GrpcChannelTest {
-    private static final GrpcTransportBuilder BUILDER = GrpcTransport.forHost("ydb.tech", 1245, "/Local");
-
-    private final AutoCloseable mocks = MockitoAnnotations.openMocks(this);
-    private final MockedStatic<ManagedChannelFactory> factoryStaticMock = Mockito.mockStatic(ManagedChannelFactory.class);
     private final ManagedChannelFactory factoryMock = Mockito.mock(ManagedChannelFactory.class);
+    private final GrpcTransportBuilder builderMock = Mockito.mock(GrpcTransportBuilder.class);
 
     @Before
     public void setUp() {
-        factoryStaticMock.when(() -> ManagedChannelFactory.fromBuilder(BUILDER)).thenReturn(factoryMock);
+        Mockito.when(builderMock.getManagedChannelFactory()).thenReturn(factoryMock);
         Mockito.when(factoryMock.getConnectTimeoutMs()).thenReturn(500l); // timeout for ready watcher
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        factoryStaticMock.close();
-        mocks.close();
     }
 
     @Test
